# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Chariot Financial Solutions case management API
servers:
  - url: /api
...
paths:
...
  /clients/bulk:
    post:
      operationId: bulkUpdateClients
      tags: [clients]
      description: Reassign several clients at once.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/BulkClientUpdate"}
      responses:
        "200":
          description: Updated clients
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/Client"}
  /clients/{id}/timeline:
    get:
      operationId: getClientTimeline
      tags: [clients]
      description: Everything that happened on this relationship, newest first - activities, logged interactions, tasks, documents, cases and portal emails.
      parameters:
        - {$ref: "#/components/parameters/Id"}
      responses:
        "200":
          description: Timeline
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/ClientTimelineItem"}
  /clients/{id}/interactions:
    get:
      operationId: listClientInteractions
      tags: [clients]
      parameters:
        - {$ref: "#/components/parameters/Id"}
      responses:
        "200":
          description: Logged interactions, newest first
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/ClientInteraction"}
    post:
      operationId: createClientInteraction
      tags: [clients]
      description: Log a call, email, meeting or note. Updates the client's last-contacted time and, when given, the next follow-up date.
      parameters:
        - {$ref: "#/components/parameters/Id"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ClientInteractionInput"}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ClientInteraction"}
  /clients/{id}/interactions/{interactionId}:
    patch:
      operationId: updateClientInteraction
      tags: [clients]
      parameters:
        - {$ref: "#/components/parameters/Id"}
        - {name: interactionId, in: path, required: true, schema: {type: integer}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ClientInteractionUpdate"}
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ClientInteraction"}
        "403":
          description: Only the author or an admin can edit an interaction
    delete:
      operationId: deleteClientInteraction
      tags: [clients]
      parameters:
        - {$ref: "#/components/parameters/Id"}
        - {name: interactionId, in: path, required: true, schema: {type: integer}}
      responses:
        "204":
          description: Deleted
        "403":
          description: Only the author or an admin can delete an interaction
...
  /documents:
    get:
      operationId: listDocuments
      tags: [clients]
      parameters:
        - {name: clientId, in: query, schema: {type: integer}}
        - {name: caseId, in: query, schema: {type: integer}}
        - {name: propertyId, in: query, schema: {type: integer}}
        - {name: category, in: query, schema: {type: string}}
        - {name: status, in: query, schema: {type: string}}
      responses: {"200": {description: Documents, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/Document"}}}}}}
    post:
      operationId: createDocumentMetadata
      tags: [clients]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/DocumentInput"}}}}
      responses: {"201": {description: Created, content: {application/json: {schema: {$ref: "#/components/schemas/Document"}}}}}
...
  /documents/{id}:
    patch:
      operationId: updateDocument
      tags: [clients]
      parameters: [{$ref: "#/components/parameters/Id"}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/DocumentUpdate"}}}}
      responses: {"200": {description: Updated, content: {application/json: {schema: {$ref: "#/components/schemas/Document"}}}}}
    delete:
      operationId: deleteDocument
      tags: [clients]
      parameters: [{$ref: "#/components/parameters/Id"}]
      responses: {"204": {description: Deleted}}
  /documents/{id}/download:
    get:
      operationId: downloadDocument
      tags: [clients]
      parameters: [{$ref: "#/components/parameters/Id"}]
      responses: {"200": {description: Document bytes}}
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: {type: integer}
...
  schemas:
...
    ClientLifecycle:
      type: string
      enum: [enquiry, onboarding, active, declined, lost]
...
    ClientAssignee:
      type: object
      required: [id, displayName]
      properties:
        id: {type: integer}
        displayName: {type: string}
    WelcomeDelivery:
      type: object
      required: [status]
      properties:
        status: {type: string, enum: [pending, sent, disabled, failed]}
        error: {type: string, nullable: true}
        deliveredAt: {type: string, format: date-time, nullable: true}
    Client:
      type: object
      required: [id, name, email, phone, companyName, onboardingStatus, createdAt, lifecycle, stale, enquiryReceivedAt, openCases, loanTotal, openTasks, termsOfBusiness, onboardingCompletedAt]
      properties:
        id: {type: integer}
        name: {type: string}
        email: {type: string}
        phone: {type: string}
        companyName: {type: string}
        onboardingStatus: {type: string, enum: [not_started, in_progress, complete]}
        createdAt: {type: string, format: date-time}
        lifecycle: {$ref: "#/components/schemas/ClientLifecycle"}
        source: {type: string, enum: [email, phone, website, referral, introducer, existing_client, other], nullable: true}
        introducerName: {type: string, nullable: true}
        introducerContact: {type: string, nullable: true}
        assignee:
          allOf: [{$ref: "#/components/schemas/ClientAssignee"}]
          nullable: true
        enquiryType: {type: string, enum: [purchase, remortgage, refinance, bridging, development, commercial, other], nullable: true}
        enquirySummary: {type: string, nullable: true}
        enquiryTimescale: {type: string, nullable: true}
        enquiryEmailText: {type: string, nullable: true}
        enquiryEmailSubject: {type: string, nullable: true}
        enquiryEmailFrom: {type: string, nullable: true}
        enquiryReceivedAt: {type: string, format: date-time}
        acceptedAt: {type: string, format: date-time, nullable: true}
        closedAt: {type: string, format: date-time, nullable: true}
        outcomeReason: {type: string, nullable: true}
        stale:
          type: boolean
          description: True for an enquiry that has waited for acceptance longer than the stale threshold.
        welcomeDelivery:
          allOf: [{$ref: "#/components/schemas/WelcomeDelivery"}]
          nullable: true
        termsOfBusiness:
          allOf: [{$ref: "#/components/schemas/TermsAcceptance"}]
          nullable: true
        onboardingCompletedAt: {type: string, format: date-time, nullable: true}
        openCases: {type: integer, description: Cases that are not completed or archived.}
        loanTotal: {type: number, description: Sum of loan amounts across open cases.}
        openTasks: {type: integer}
        lastActivityAt: {type: string, format: date-time, nullable: true}
        lastContactedAt: {type: string, format: date-time, nullable: true, description: Most recent logged interaction.}
        nextFollowUpAt: {type: string, pattern: '^\d{4}-\d{2}-\d{2}$', nullable: true}
        title: {type: string, nullable: true}
        dateOfBirth: {type: string, pattern: '^\d{4}-\d{2}-\d{2}$', nullable: true}
        nationality: {type: string, nullable: true}
        maritalStatus: {type: string, nullable: true}
        dependants: {type: integer, nullable: true}
        currentAddress: {type: string, nullable: true}
        currentAddressCity: {type: string, nullable: true}
        currentAddressPostcode: {type: string, nullable: true}
        previousAddress: {type: string, nullable: true}
        previousAddressCity: {type: string, nullable: true}
        previousAddressPostcode: {type: string, nullable: true}
        alternativePhone: {type: string, nullable: true}
        employmentStatus: {type: string, nullable: true}
        employerName: {type: string, nullable: true}
        jobTitle: {type: string, nullable: true}
        annualIncome: {type: number, nullable: true}
        otherIncome: {type: number, nullable: true}
        monthlyCommitments: {type: number, nullable: true}
        creditHistoryNotes: {type: string, nullable: true}
        companyNumber: {type: string, nullable: true}
        companyRegisteredAddress: {type: string, nullable: true}
        companyRegisteredCity: {type: string, nullable: true}
        companyRegisteredPostcode: {type: string, nullable: true}
        notes: {type: string, nullable: true}
...
    BulkClientUpdate:
      type: object
      required: [ids]
      properties:
        ids: {type: array, items: {type: integer}, minItems: 1, maxItems: 200}
        assignedUserId: {type: integer, nullable: true}
    InteractionKind:
      type: string
      enum: [call, email, meeting, note]
    ClientInteraction:
      type: object
      required: [id, clientId, caseId, kind, summary, occurredAt, createdByUserId, createdByName, createdAt]
      properties:
        id: {type: integer}
        clientId: {type: integer}
        caseId: {type: integer, nullable: true}
        kind: {$ref: "#/components/schemas/InteractionKind"}
        summary: {type: string}
        occurredAt: {type: string, format: date-time}
        createdByUserId: {type: integer, nullable: true}
        createdByName: {type: string}
        createdAt: {type: string, format: date-time}
    ClientInteractionInput:
      type: object
      required: [kind, summary]
      properties:
        kind: {$ref: "#/components/schemas/InteractionKind"}
        summary: {type: string, minLength: 1, maxLength: 4000}
        occurredAt: {type: string, format: date-time}
        caseId: {type: integer, nullable: true}
        nextFollowUpAt:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
          nullable: true
          description: Sets the client's next follow-up date. Omit to leave it unchanged; null clears it.
    ClientInteractionUpdate:
      type: object
      properties:
        kind: {$ref: "#/components/schemas/InteractionKind"}
        summary: {type: string, minLength: 1, maxLength: 4000}
    ClientTimelineItem:
      type: object
      required: [id, kind, title, detail, actorName, occurredAt, href]
      properties:
        id: {type: string}
        kind: {type: string, enum: [activity, interaction, task, document, case, email]}
        title: {type: string}
        detail: {type: string}
        actorName: {type: string}
        occurredAt: {type: string, format: date-time}
        href: {type: string, nullable: true}
        interactionKind:
          allOf: [{$ref: "#/components/schemas/InteractionKind"}]
          nullable: true
...
    TermsAcceptance:
      type: object
      required: [acceptedAt, via, version, note, acceptedBy]
      properties:
        acceptedAt: {type: string, format: date-time}
        via: {type: string, enum: [portal, signed_upload, staff]}
        version: {type: integer, nullable: true}
        note: {type: string, nullable: true}
        acceptedBy: {type: string, nullable: true, description: Staff member who recorded it, when not accepted by the client in the portal.}
...
    Document:
      type: object
      required: [id, name, category, status]
      properties:
        id: {type: integer}
        name: {type: string}
        category: {type: string}
        status: {type: string}
        clientId: {type: integer}
        caseId: {type: ["integer", "null"]}
        contentType: {type: ["string", "null"]}
        byteSize: {type: ["integer", "null"]}
        uploadedAt: {type: ["string", "null"], format: date-time}
    DocumentInput:
      type: object
      required: [clientId, name, category]
      properties:
        clientId: {type: integer}
        caseId: {type: ["integer", "null"]}
        name: {type: string, minLength: 1, maxLength: 255}
        category: {type: string, minLength: 1}
        status: {type: string}
    DocumentUpdate:
      type: object
      properties:
        category: {type: string, minLength: 1}
        status: {type: string, minLength: 1}
...